# wallet2, RPC: more updates to optimize requests for pool txs