- `/getblocks.bin` respects the `RESTRICTED_TX_COUNT` (=100) when returning pool txs via a restricted RPC daemon.
	- A restricted RPC daemon includes a max of `RESTRICTED_TX_COUNT` txs in the `added_pool_txs` field, and returns any remaining pool tx hashes in the `remaining_added_pool_txids` field. The client then requests the remaining txs via `/gettransactions` in chunks.
- `/gettransactions` no longer does expensive no-ops over ALL pool txs if the client requests a subset of pool txs. Instead it searches for the txs the client explicitly requests.
	- Users will  benefit from this even if they continue using a wallet that **does not** use this PR, but point to a daemon that **does** use this PR.
- Reset `m_pool_info_query_time` when a user:
	- rescans the chain (so that the wallet re-requests the whole pool)
	- changes the daemon their wallet points to (a new daemon would have a different view of the pool)
- `/getblocks.bin` respects the `req.prune` field when returning pool txs.
- Pool extension fields in the response to `/getblocks.bin` are now optional with default 0'd values.

________

Benchmarks when requesting blocks and pool txs from a daemon with restricted RPC (okay-ish core i7, 32gb RAM, NVMe SSD):


|					          |    # of txs in pool  |   # of trips to daemon  |  Time daemon spent processing requests |  % speedup
| :-- | --:  | --: | --: | :--: |
| | | | | |
v0.18.1.2           |       101 small      |           4 					  | 					       113ms							    |
8076 + this		|       101 small	     |           2 					  |  								41ms    			        |    *64%*
| | | | | |
| | | | | |
v0.18.1.2	          | 	   1001 small      |          13						  |	              3432ms			            |
8076 + this		| 	   1001 small      |          11						  |                 438ms								  |		 *87%*
| | | | | |
| | | | | |
v0.18.1.2           |       101 large      |           4 					  | 					      1786ms							    |
8076 + this		|       101 large	     |           2 					  |  							 436ms    			        |    *76%*
| | | | | |
| | | | | |
v0.18.1.2	          |       301 large	     |           6						  |	              9320ms  			          |
8076 + this		|       301 large	     | 	        4						  |	              1007ms								  |		 *89%*

Large tx = ~141   kb
Small tx =   ~1.5 kb

*Note*: in v0.18.1.2, the daemon spends the majority of its time processing `/gettransactions` requests, doing expensive no-ops for all pool txs. Since daemons running this PR will no longer do these no-ops, all v0.18 wallets that point to daemons running this PR will see significant speedups querying the pool when the network is experiencing higher load.

Update 12/15/22: when pointing to a restricted RPC daemon, `/get_pool_info`'s non-incremental path doesn't need to query the db *two* times for all pool txs, just *once* is fine. This reduces the time the daemon spends processing the first request in the benchmarks above.