@rbrunner7 2022-12-18
First of all, many thanks for this PR. I don't think that I would have found the energy to yet again go over my code and make such extensive chances.

I clearly see that if you have a limit about how many transactions at once a restricted daemon is allowed to give back, we have to enforce that limit. But for me it's a pity how much code that needs on top of the already quite extensive code for incremental pool fetching, how the added code makes everything even more complex still, and how far away this PR will push the `wallet2` code move away from status quo ...