@rbrunner7 2022-12-18
I tested your code in the following scenario:

I went down with `RESTRICTED_TRANSACTIONS_COUNT` from 100 to 2 in all three relevant places, to make testing of that "fetch additional transactions beyond the limit" much easier. Then I set up a restricted daemon to make that limit active.

I used testnet and made sure that with each new block immediately 4 new transactions get submitted, so the pool is almost always filled with 4 transactions. (I used my TechWallet for that which you can instruct to repeat transactions over the next 10 blocks. Initiate that 4 times, using of course a wallet with the necessary number of unspent outputs, and you achieve the described insta-filling of the pool.)

I then ran an instance of `monero-wallet-cli` with any testnet wallet, compiled to log to console for easier watching.

I saw the following behavior. I did not try to debug, because it's my hope that when you read this you have an almost instant revelation where the problem is and can correct right away. If not, I will try to debug myself :)

With a high limit for `RESTRICTED_TRANSACTIONS_COUNT` everything works as expected: If I just let the wallet stand and background-sync it sees new pool transactions in chunks of 4 whenever a new block arrives.

However, with a limit of 2 for `RESTRICTED_TRANSACTIONS_COUNT` it lists chuncks of **6** pool transactions as new, where the 2 first and the 2 next transactions are identical. Here a relevant snippet from the log:

````
2022-12-18 09:45:31.712	T update_pool_state_from_pool_data start
2022-12-18 09:45:31.712	I Found new pool tx: <05341894c00569b0f563a8fa82277c756d178927ac183cbb7a10e04391d53e21>
2022-12-18 09:45:31.712	I Found new pool tx: <45654e702dbfa75872235d678a85d114208d049c778e6d897d43ddff9c33ff46>
2022-12-18 09:45:31.712	I Found new pool tx: <05341894c00569b0f563a8fa82277c756d178927ac183cbb7a10e04391d53e21>
2022-12-18 09:45:31.712	I Found new pool tx: <45654e702dbfa75872235d678a85d114208d049c778e6d897d43ddff9c33ff46>
2022-12-18 09:45:31.712	I Found new pool tx: <34953d20d8a3ce452fc2b5c16d78681d634ff10b740caaf3da446cbdec049895>
2022-12-18 09:45:31.713	I Found new pool tx: <f29ef26b7e54cca8a112222da3ff0823506231d2cd49c4b693b297a9dc4827c4>
2022-12-18 09:45:31.713	T update_pool_state_from_pool_data end
2022-12-18 09:45:31.713	I Refresh done, blocks received: 0, balance (all accounts): 28.494490151099, unlocked: 28.494490151099
````

Nothing goes really wrong of course, but you can somehow suspect that this "fetch additional transactions beyond the limit" does not yet work as it should.