@j-berman 2022-12-26
(squashed the fix to a restricted daemon [returning duplicate added txids](https://github.com/rbrunner7/monero/pull/4#issuecomment-1356763545) in the incremental path)