@j-berman 2023-01-21
To run benchmarks on the time it takes the daemon to process requests:

1. Get the chain set up with all txs ready to go. DM me on matrix/IRC (jberman:matrix.org) and I will send files that will help you do this.

2. Start up a daemon with the flags `--restricted-rpc --data-dir <path to chain data>`

3. Open a new wallet and sync it with the daemon, then close the wallet.

4. Submit txs to the pool.

5. Set log level 2 on the daemon and open the wallet.

6. Here's how to measure the benchmarks...

On a v0.18.1.2 daemon, check the logs to see how long the following RPCs take to process: `/get_transaction_pool_hashes` + `/gettransactions` + `/getblocks.bin`. Double check the wallet calls `/gettransactions` `round_up(# of txs in pool / 100)` times, and the other RPCs once.

On a daemon running https://github.com/monero-project/monero/pull/8076, check the logs to see how long `/getblocks.bin` + `/gettransactions`  take to process. Double check the wallet calls `/getblocks.bin` once and `/gettransactions` `round_up(# of txs in pool / 100) - 1` times.

Search the logs for "processed with" to see how long it takes. e.g. `/getblocks.bin() processed with 0/427/2ms` means it took 427ms to process.

7. Start over from step 3 for each test.