@plowsof 2023-02-11
will update these benchmarks with speed ups after figuring some things out:
not okish i5 nvme ssd 8GB ram
| | # of tx in pool | # trips to daemon |  Time daemon spent processing requests | |
| --- | --- | --- | --- | --- |
| 8076 + this | 101 small | 2  | 79 | |
| 8076 + this | 1001 small | 11  | 456 | |
| 8076 + this | 101 large | 2  | 387 | |
| 8076 + this |301 large | 4  | 1036 | |

i need to figure out if im doing something wrong:
```
old wallet to new daemon 101 large 
/get_transaction_pool_hashes.bin processed with 0/0/0ms
/gettransactions processed with 0/387/4ms
/gettransactions processed with 0/3/0ms
/getblocks.bin() processed with 0/0/0ms
/get_transaction_pool_hashes.bin processed with 0/0/0ms
/getblocks.bin() processed with 0/0/0ms
```

```
new wallet to old daemon 101 large 
/gethashes.bin() processed with 0/3/0ms
/gethashes.bin() processed with 0/3/1ms
/gethashes.bin() processed with 0/2/1ms
/gethashes.bin() processed with 0/1/0ms
/get_transaction_pool_hashes.bin processed with 0/0/0ms
/getblocks.bin() processed with 0/1/0ms
/get_transaction_pool_hashes.bin processed with 0/0/0ms
/getblocks.bin() processed with 0/0/0ms
/getblocks.bin() processed with 0/0/0ms
/get_transaction_pool_hashes.bin processed with 0/1/0ms
/gettransactions processed with 0/975/5ms
/gettransactions processed with 0/659/0ms
/getblocks.bin() processed with 0/0/0ms
/get_transaction_pool_hashes.bin processed with 0/0/0ms
```

```
old to old 101 large (setup issues on my end i think causing multiple calls etc)
/gethashes.bin() processed with 0/2/0ms
/gethashes.bin() processed with 0/2/0ms
/gethashes.bin() processed with 0/2/0ms
/gethashes.bin() processed with 0/0/0ms
/get_transaction_pool_hashes.bin processed with 0/0/0ms
/gettransactions processed with 0/964/5ms
/gettransactions processed with 0/673/0ms
/getblocks.bin() processed with 0/0/0ms
/get_transaction_pool_hashes.bin processed with 0/0/0ms
/getblocks.bin() processed with 0/0/0ms
/get_transaction_pool_hashes.bin processed with 0/0/0ms
/getblocks.bin() processed with 0/1/0ms
/gethashes.bin() processed with 0/8/0ms
/gethashes.bin() processed with 0/8/1ms
/gethashes.bin() processed with 0/12/0ms
/gethashes.bin() processed with 0/1/0ms
/gethashes.bin() processed with 0/3/0ms
/get_transaction_pool_hashes.bin processed with 0/0/1ms
/gettransactions processed with 0/927/5ms
/gettransactions processed with 0/661/0ms
/getblocks.bin() processed with 0/0/0ms
/get_transaction_pool_hashes.bin processed with 0/0/0ms
/getblocks.bin() processed with 0/0/0ms
```

im going to continue QA testing tomorrow! (mining blocks instead of flush tx pool)